Extract settings utils to own separate file
